SONAR-12043 increase default timeout to 6 hours

+++ server/sonar-ce/src/main/java/org/sonar/ce/configuration/CeConfiguration.java
@@ -50,6 +50,6 @@ public interface CeConfiguration {
   /**
    * Delay before stopping workers during a graceful timeout using milliseconds unit.
    */
-  int getGracefulStopTimeoutInMs();
+  long getGracefulStopTimeoutInMs();
 
 }
